Tidied up conftest
