+++ tasks/db_tasks.py
@@ -59,6 +59,8 @@ def generate_test_data(c):
 
 @task(
     help={
+@task(
+    help={
         "fundround": "The round and fund to seed applications for assessment.",
         "appcount": "The amount of applications to seed.",
     }
@@ -75,7 +77,7 @@ def seed_dev_db(c, fundround=None, appcount=None):
         with app.app_context():
             from fsd_utils import CommonConfig
             from uuid import uuid4
-            from tests.conftest import seed_database_for_fund_round
+            from tests._helpers import seed_database_for_fund_round
             from config import Config
 
             config = {
@@ -91,14 +93,19 @@ def seed_dev_db(c, fundround=None, appcount=None):
                 fund_round = config[fundround]
                 apps = int(appcount)
                 print(
-                    f"Seeding {apps} applications for fund_round: "
-                    f"'{fundround}'"
+                    f"Seeding {apps} applications for "
+                    f"fund_round: '{fundround}'"
                 )
 
             while choosing:
 
                 new_line = "\n"
                 _echo_print(
+                    f"fund-rounds available to seed: {new_line} -"
+                    " {f' {new_line} - '.join(config.keys())}",
+                )
+                fund_round_input = str(
+                    _echo_input("Please type the fund-round to seed:")
                     f"fund-rounds available to seed: "
                     f"{new_line} - {f' {new_line} - '.join(config.keys())}",
                 )
@@ -109,6 +116,8 @@ def seed_dev_db(c, fundround=None, appcount=None):
                 apps = int(_echo_input("How many applications?"))
                 choosing = (
                     not _echo_input(
+                        f"Would you like to insert {apps} applications"
+                        " for {fund_round_input}? y/n \n"
                         f"Would you like to insert {apps} applications for "
                         f"{fund_round_input}? y/n \n"
                     ).lower()
